+++ src/Blackjack/PlayerControls.cpp
@@ -18,19 +18,19 @@ PlayerControls::PlayerControls()
     DoubleButton = this->MakeSprite("DoubleButton");
 
 
-    int baseYPos = 25;
+    float baseYPos = 25;
 
     DrawButton->SetTexture("assets//buttons//DrawButton.png");
-    DrawButton->SetPos(baseYPos, 0);
+    DrawButton->SetPos(baseYPos, 0.0);
 
     HoldButton->SetTexture("assets//buttons//HoldButton.png");
-    HoldButton->SetPos(155 * 1 + baseYPos, 0);
+    HoldButton->SetPos(155 * 1 + baseYPos, 0.0);
 
     SplitButton->SetTexture("assets//buttons//SplitButton.png");
-    SplitButton->SetPos(155 * 2 + baseYPos, 0);
+    SplitButton->SetPos(155 * 2 + baseYPos, 0.0);
 
     DoubleButton->SetTexture("assets//buttons//DoubleButton.png");
-    DoubleButton->SetPos(155 *3 + baseYPos, 0);
+    DoubleButton->SetPos(155 *3 + baseYPos, 0.0);
 }
 
 PlayerControls::~PlayerControls()
